+++ molo/profiles/admin.py
@@ -5,10 +5,10 @@ from django.http import HttpResponse
 from django.contrib.auth.models import User
 from django.contrib.auth.admin import UserAdmin
 from django.contrib.admin.sites import NotRegistered
+from django.utils.translation import ugettext_lazy as _
 
 from daterange_filter.filter import DateRangeFilter
-from wagtailmodeladmin.options import ModelAdmin as WagtailModelAdmin
-
+from wagtail.contrib.modeladmin.options import ModelAdmin as WagtailModelAdmin
 from molo.profiles.admin_views import FrontendUsersAdminView
 
 try:
@@ -63,26 +63,50 @@ class ProfileUserAdmin(UserAdmin):
             return obj.profile.date_of_birth
         return ''
 
+    def _site(self, obj, *args, **kwargs):
+        if hasattr(obj, 'profile') and obj.profile.site:
+            return obj.profile.site
+        return ''
+
 
 # Below here is for Wagtail Admin
 class FrontendUsersDateRangeFilter(DateRangeFilter):
     template = 'admin/frontend_users_date_range_filter.html'
 
 
+class CustomUsersListFilter(admin.SimpleListFilter):
+    title = _('Type')
+    parameter_name = 'usertype'
+
+    def lookups(self, request, model_admin):
+        return (
+            ('frontend', _('Frontend Users')),
+            ('admin', _('Admin Users')),
+        )
+
+    def queryset(self, request, queryset):
+        if self.value() == 'frontend':
+            return queryset.filter(is_staff=False, groups__isnull=True)
+
+        if self.value() == 'admin':
+            return queryset.exclude(is_staff=False, groups__isnull=True)
+
+
 class FrontendUsersModelAdmin(WagtailModelAdmin, ProfileUserAdmin):
     model = User
-    menu_label = 'End Users'
+    menu_label = 'Users Export'
     menu_icon = 'user'
     menu_order = 600
     index_view_class = FrontendUsersAdminView
     add_to_settings_menu = True
     list_display = ('username', '_alias', '_mobile_number', '_date_of_birth',
-                    'email', 'date_joined', 'is_active')
+                    'email', 'date_joined', 'is_active', '_site')
 
-    list_filter = (('date_joined', FrontendUsersDateRangeFilter), 'is_active')
+    list_filter = (
+        ('date_joined', FrontendUsersDateRangeFilter), 'is_active',
+        CustomUsersListFilter)
 
     search_fields = ('username',)
 
     def get_queryset(self, request):
-        queryset = User.objects.filter(is_staff=False)
-        return queryset
+        return User.objects.filter(profile__site=request.site)