+++ CHANGES.rst
@@ -1,5 +1,65 @@
 CHANGE LOG
 ==========
+4.0.1-beta.2
+------------
+- Link users to Site
+- Add multisite functionality to user profiles
+
+4.0.0-beta.2
+------------
+- Pre-release to support molo v5
+
+3.0.0
+-----
+- Add terms and conditions on registration
+
+2.2.1
+-----
+- Updated minimum molo core dependency to 4.3.2
+
+2.2.0
+-----
+- Ensure that security question index page cannot be deleted
+
+2.1.3
+-----
+- Fix bug that allowed user's mobile number and email to be deleted when updated despite being required fields
+
+2.1.2
+-----
+- Use better python logic in forms
+
+2.1.1
+-----
+- Fixed bug forcing users to add mobile number when not required
+
+2.1.0
+-----
+- Update template to only show mobile number field if it's activated and has country calling code
+- Show a warning message on CMS if mobile number is activated but country calling code has not been set
+
+2.0.2
+-----
+- Added missing button class for templates
+
+2.0.1
+-----
+- Updated templates in order to reflect styling changes in modeladmin
+
+2.0.0
+-----
+- Removed dependency on wagtailmodeladmin
+
+Backwards incompatible changes
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+- Removed use of ``wagtailmodeladmin``: use ``wagtail.contrib.modeladmin`` instead
+- ``{% load wagtailmodeladmin_tags %}`` has been replaced by ``{% load modeladmin_tags %}``
+
+NOTE: This release is not compatible with molo versions that are less than 4.0
+
+1.5.1
+-----
+- Fix duplicated users via wagtail admin
 
 1.5.0
 -----